Setup experiment 01
Experiment 01 is mostly the original one, but with the data expanded to
include the characterization of the categories (for example: for this
experiment, "plant", and "tree", and "flower" are labels too,
rather than only the dendrogram's leaves (lowermost "categories"
i.e. "pine", "oak", "daisy", "rose", etc)

Data for this experiment is in data/data_01/data.csv
This is essentially the only meaningful change. Some change in
analysis/main's dendrogram function to display only the lowermost
labels, but more change would be needed (i.e. not all code has been
adapted)