Setup experiment 02
Lifelong learning setup: now we withhold part of the training set, and
release it gradually. First the network is only given "animal" and
"plant" as items. Then the intermediary categories as well, and finally,
all items.

The data file is the same as the previous one, but when creating the
Dataset object, only some of the dataset items are included.